clean up unused file
