+++ .rspec
@@ -1,6 +1,4 @@
 --require spec_helper.rb
 --require tic_tac_toe_rspec.rb
---require player_rspec.rb
---require game_rspec.rb
